+++ Application/Categories/Queries/Common/CategoryDetailVM.cs
@@ -1,23 +1,20 @@
 ﻿using AutoMapper;
+using SimpleCMS.Application.Categories.Queries.GetCategoryDetail;
 using SimpleCMS.Application.Common.Mappings;
 using SimpleCMS.Domain.Entities;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace SimpleCMS.Application.Categories.Queries.GetCategoryDetail
+namespace SimpleCMS.Application.Categories.Queries.Common
 {
     public class CategoryDetailVM : IMapFrom<Category>
     {
         public int Id { get; set; }
 
-        public string Name { get; set; }
+        public string Title { get; set; }
 
         public string Description { get; set; }
 
-        public virtual List<CategoryTopicDTO> Topics { get; set; }
+        public virtual List<CategoryTopicVM> Topics { get; set; }
 
         public void Mapping(Profile profile) =>
             profile.CreateMap<Category, CategoryDetailVM>()