+++ Application/Categories/Queries/GetCategoriesList/CategoriesListVM.cs
@@ -1,10 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using SimpleCMS.Application.Categories.Queries.Common;
+using System.Collections.Generic;
 
 namespace SimpleCMS.Application.Categories.Queries.GetCategoriesList
 {
     public class CategoriesListVM
     {
-        public IList<CategoryDTO> Categories { get; set; }
+        public IList<CategoryDetailVM> Categories { get; set; }
         public int Count { get; set; }
     }
 }