+++ Application/Categories/Queries/GetCategoryDetail/CategoryTopicVM.cs
@@ -4,16 +4,16 @@ using SimpleCMS.Domain.Entities;
 
 namespace SimpleCMS.Application.Categories.Queries.GetCategoryDetail
 {
-    public class CategoryTopicDTO : IMapFrom<Topic>
+    public class CategoryTopicVM : IMapFrom<Topic>
     {
         public int Id { get; set; }
 
-        public string Name { get; set; }
+        public string Title { get; set; }
         public string Description { get; set; }
 
 
         public void Mapping(Profile profile) =>
-            profile.CreateMap<Topic, CategoryTopicDTO>()
+            profile.CreateMap<Topic, CategoryTopicVM>()
             .ForMember(d => d.Id, opt => opt.MapFrom(e => e.TopicId));
     }
 }