+++ Application/Categories/Queries/GetCategoryDetail/GetCategoryDetailQueryHandler.cs
@@ -2,11 +2,9 @@
 using AutoMapper.QueryableExtensions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
+using SimpleCMS.Application.Categories.Queries.Common;
 using SimpleCMS.Application.Common.Interfaces;
-using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;