+++ Application/Topics/Queries/GetTopicDetail/GetTopicDetailQuery.cs
@@ -1,9 +1,5 @@
 ﻿using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using SimpleCMS.Application.Topics.Queries.Common;
 
 namespace SimpleCMS.Application.Topics.Queries.GetTopicDetail
 {