+++ Application/Topics/Queries/GetTopicsList/GetTopicsListQueryHandler.cs
@@ -1,11 +1,11 @@
-﻿using AutoMapper.QueryableExtensions;
+﻿using AutoMapper;
+using AutoMapper.QueryableExtensions;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using SimpleCMS.Application.Common.Interfaces;
+using SimpleCMS.Application.Topics.Queries.Common;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Linq;
-using Microsoft.EntityFrameworkCore;
-using AutoMapper;
 
 namespace SimpleCMS.Application.Topics.Queries.GetTopicsList
 {
@@ -23,7 +23,7 @@ namespace SimpleCMS.Application.Topics.Queries.GetTopicsList
         public async Task<TopicsListVM> Handle(GetTopicsListQuery request, CancellationToken cancellationToken)
         {
             var topics = await _context.Topics
-                .ProjectTo<TopicDTO>(_mapper.ConfigurationProvider)
+                .ProjectTo<TopicDetailVM>(_mapper.ConfigurationProvider)
                 .ToListAsync();
 
             var viewModel = new TopicsListVM()